+++ docs/src/code/core/utils.rst
@@ -6,6 +6,7 @@ Utilities
    :caption: Utility modules
 
    utils/format_trials
+   utils/format_terminal
 
 .. automodule:: orion.core.utils
    :members:
