+++ docs/src/code/core/utils/format_terminal.rst
@@ -1,0 +1,5 @@
+Format terminal
+===============
+
+.. automodule:: orion.core.utils.format_terminal
+   :members:
