+++ tests/unittests/core/cli/test_info.py
@@ -5,11 +5,11 @@ import itertools
 
 import pytest
 
-from orion.core.cli.info import (
+from orion.core.io.space_builder import SpaceBuilder
+from orion.core.utils.format_terminal import (
     format_algorithm, format_commandline, format_config, format_dict, format_identification,
     format_info, format_list, format_metadata, format_refers, format_space, format_stats,
     format_title, get_trial_params)
-from orion.core.io.space_builder import SpaceBuilder
 from orion.core.worker.trial import Trial
